requirements.txt
prettytable
pyyaml
ynab==0.0.3